qcow2-rs: fix(meta): Prevent panic on malformed qcow2 header
A malformed qcow2 header with an invalid cluster_bits value could
cause a panic due to a left-shift overflow when calculating the
cluster size. This could be triggered by a crafted qcow2 image,
leading to a denial of service.

This commit fixes the issue by validating the cluster_bits field
before it is used. The value is now checked to be within a valid
range (9 to 30, inclusive) to prevent the overflow. If the value is
out of range, the header parsing function now returns an error instead
of panicking.

A test case with a known malformed header has been added to ensure
this vulnerability does not regress.

Signed-off-by: Ming Lei <ming.lei@redhat.com>